newtype uint64 = fromNum nat
Replace the entire machine numbers model with a single type, uint64,
which is a wrapper around a nat. To retain some level of soundness,
during extraction we'll do overflow checks and throw an exception (so at
least the program doesn't execute doing something wrong).

It does seem like uint64s are good enough for our purposes; uint32s are
fairly similar in performance and size, and the one use case for uint16s
was for encoding value lengths. If we want that back, we can always
provide a function to encode/decode a bounded u64 in two bytes.

+++ src/Database/Common.v
@@ -13,5 +13,5 @@ End Entry.
 Module SliceHandle.
   Record t :=
     mk { offset : uint64;
-         length : uint32; }.
+         length : uint64; }.
 End SliceHandle.
